requirements.txt
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
get==2019.4.13
post==2019.4.13
psycopg2==2.8.4
request==2019.4.13
requests==2.22.0
urllib3==1.25.8
django-rest-swagger=2.2.0